JHistint.jl - Julia Histopathology Interface

Julia interface for implementing the REST APIs available on the Cancer Slide Digital Archive (CDSA) portal for downloading histological images available in The Cancer Genome Atlas (TCGA). The Cancer Slide Digital Archive (CDSA) is a web platform for support, sharing, and analysis of digital pathological data. Currently, it hosts over 23,000 images associated with the data available on "The Cancer Genome Atlas" Data Portal. The library includes functions for managing image-processing algorithms for cell segmentation, constructing the adjacency matrix, and interfacing with the J-Space.jl package.

Link GitHub repository: JHistint.jl

Link GitHub repository: J-Space.jl

CDSA Portal: Clicca qui

Repository containing the data mapped in the portal: Clicca qui

Guide to using the APIs: Clicca qui

Package Structure

The case and collection folders store metadata in .json format for individual cases and collections available on the TCGA Data Portal. The collection folder is structured as follows:

collectionlist.json = Stores access data (metadata) for collections (Projects in TCGA).

collection_name.json = Stores access data (metadata) for a single collection. The .json file is generated based on the collection chosen by the user.

The case folder is structured as follows:

collection_name.json = Stores all metadata related to cases associated with the collection selected by the user.

The slides folder stores histological images related to individual cases. The images are organized based on collection (TCGA-chol, TCGA-esca, etc.), and the individual case being analyzed (TCGA-2H-A9GF, TCGA-2H-A9GG, etc.). Within each folder related to the case, the slides are stored in compressed .zip files. The format of each individual slide is .tif. The folder names related to the cases correspond to the values of the Case ID field listed in the TCGA Data Portal. The names of the .zip files located in each folder refer to the Sample ID attribute associated with the patient. The slide name is given by concatenating the Slide ID and Slide UUID attributes that can be found in the lower section of the web page dedicated to the generic case TCGA-XX-YYYY.

Example: TCGA-02-0001-01C-01-TS1.zip  
  - 02 = refers to the TSS (Tissue Source Site).  
  - 0001 = refers to the code associated with the Participant, an alphanumeric string.  
  - 01 = refers to the Sample Type. The values associated with tumor samples are in the range 01-09. 10-19 indicates the range for non-diseased normal samples. 20-29 indicates samples currently under control.  
  - C = refers to the Vial field related to the ordering of the sample in the sample sequence. Values range from A-Z.  
  - 01 = refers to the Portion field related to the ordering of the analyzed portions associated with a sample. It takes values in the range 01-99.  
  - TS1 = refers to the Slide field related to the type of image. The values that can be assumed are TS (Top Slide), BS (Bottom Slide), and MS (Middle Slide). The alphanumeric value indicates the slide ordering.

JHistint Collections

The available collections are:

TCGA-BRCA = Breast Invasive Carcinoma (Breast)

TCGA-OV = Ovarian Serous Cystadenocarcinoma (Ovary)

TCGA-LUAD = Lung Adenocarcinoma (Bronchus and Lung)

TCGA-UCEC = Uterine Corpus Endometrial Carcinoma (Corpus uteri)

TCGA-GBM = Glioblastoma Multiforme (Brain)

TCGA-HSNC = Head and Neck Squamous Cell Carcinoma (Larynx, Lip, Tonsil, Gum, Other and unspecified parths of mouth)

TCGA-KIRC = Kidney Renal Clear Cell Carcinoma (Kidney)

TCGA-LGG = Brain Lower Grade Glioma (Brain)

TCGA-LUSC = Lung Squamous Cell Carcinoma (Bronchus and lung)

TCGA-TCHA = Thyroid Carcinoma (Thyroid gland)

TCGA-PRAD = Prostate Adenocarcinoma (Prostate gland)

TCGA-SKCM = Skin Cutaneous Melanoma (Skin)

TCGA-COAD = Colon Adenocarcinoma (Colon)

TCGA-STAD = Stomach Adenocarcinoma (Stomach)

TCGA-BLCA = Bladder Urothelial Carcinoma (Bladder)

TCGA-LIHC = Liver Hepatocellular Carcinoma (Liver and intrahepatic bile ducts)

TCGA-CESC = Cervical Squamous Cell Carcinoma and Endocervical Adenocarcinoma (Cervix uteri)

TCGA-KIRP = Kidney Renal Papillary Cell Carcinoma (Kidney)

TCGA-SARC = Sarcoma (Various)

TCGA-ESCA = Esophageal Carcinoma (Esophagus)

TCGA-PAAD = Pancreatic Adenocarcinoma (Pancreas)

TCGA-READ = Rectum Adenocarcinoma (Rectum)

TCGA-PCPG = Pheochromocytoma and Paraganglioma (Adrenal gland)

TCGA-TGCT = Testicular Germ Cell Tumors (Testis)

TCGA-THYM = Thymoma (Thymus)

TCGA-ACC = Adrenocortical Carcinoma -Adenomas and Adenocarcinomas (Adrenal gland)

TCGA-MESO = Mesothelioma (Heart, mediastinum and pleura)

TCGA-UVM = Uveal Melanoma (Eye and adnexa)

TCGA-KICH = Kidney Chromophobe (Kidney)

TCGA-UCS = Uterine Carcinosarcoma (Uterus, NOS)

TCGA-CHOL = Cholangiocarcinoma (Liver and intrahepatic bile ducts, Other and unspecified part of biliary track)

TCGA-DLBC = Lymphoid Neoplasm Diffuse Large B-cell Lymphoma (Various)

To download a specific collection, just indicate the name of the collection: BRCA, OV, LUAD.

Package Installation

The JHistint package is available in the Julia Registries and can be installed as follows:

julia > using Pkg
julia > Pkg.add("JHistint")
julia > using JHistint

Otherwise, type ] in the Julia REPL and execute:

(@v1.8) pkg > add JHistint
(@v1.8) pkg > using JHistint

Download Slides functions (JHistint.jl)

JHistint.download_single_collection — Method

download_single_collection(collection_name::AbstractString)

Function for downloading histological slides associated with a collection available in TCGA.

Arguments

collection_name::AbstractString = Collection of TCGA data to download the histological slides.

Notes

The function evaluates the collection_name argument, and in case of an invalid collection, considers the configuration in the Config.toml file. The value set in the package is default.

# Examples with valid input
julia> JHistint.download_single_collection("acc")
julia> JHistint.download_single_collection("bLca")

# Examples with invalid input
julia> JHistint.download_single_collection("ac")
julia> JHistint.download_single_collection("")

JHistint.download_all_collection — Method

download_all_collection()

Function for downloading histological slides associated with all collections available in TCGA.

# Examples with valid input
julia> JHistint.download_all_collection()

Cell Segmentation Slides functions (JHistint.jl)

JHistint.slide_cell_segmentation_without_download — Method

slide_cell_segmentation_without_download(collection_name::AbstractString)

Function for performing cell segmentation on histopathological slides present in the JHistint_DB database associated with the collection name provided as an argument. After generating the segmented slide, the function proceeds with constructing and saving the corresponding graph and adjacency matrix.

Arguments

collection_name::AbstractString = Collection of TCGA data to download the histological slides.

Notes

The function utilizes the JHistint_DB database for performing cell segmentation on the histopathological slides associated with the provided collection name. It generates a segmented slide and constructs a corresponding graph and adjacency matrix. The output files are saved in a user-defined directory. The function may take a considerable amount of time to complete, depending on the size of the slides and the complexity of the segmentation algorithm. For each slide in the database, cell segmentation is performed using the apply_segmentation_without_download function, and the path where the result is saved is stored in the database using the load_seg_slide function. The segmentation process is defined in 4 steps:

LOAD SLIDE ... (slide_id)

APPLY SEGMENTATION ... (slide_id)

BUILD GRAPH ... (slide_id)

BUILD & SAVE ADJACENCY MATRIX ... (slide_id)

J-SPACE features ... (slide_id)

The adjacency matrix is saved in the same directory as the original image in text format. Finally, a confirmation message is printed for each segmented slide. Unlike the slide_cell_segmentation_with_download function, this function does not involve the creation and download of the segmented image.

# Examples with valid input
julia> JHistint.slide_cell_segmentation_without_download("acc")
julia> JHistint.slide_cell_segmentation_without_download("bLca")

# Examples with invalid input
julia> JHistint.slide_cell_segmentation_without_download("ac")
julia> JHistint.slide_cell_segmentation_without_download("")

JHistint.slide_cell_segmentation_with_download — Method

slide_cell_segmentation_with_download(collection_name::AbstractString)

Function for performing cell segmentation on histopathological slides present in the JHistint_DB database associated with the collection name provided as an argument. The function downloads the segmented slide, which is placed in the same directory as the original slide. After generating the segmented slide, the function proceeds with constructing and saving the corresponding graph and adjacency matrix.

Arguments

collection_name::AbstractString = TCGA data collection for which to perform cell segmentation.

Notes

The function utilizes the JHistint_DB database for performing cell segmentation on the histopathological slides associated with the provided collection name. It generates a segmented slide and constructs a corresponding graph and adjacency matrix. The output files are saved in a user-defined directory. The function may take a considerable amount of time to complete, depending on the size of the slides and the complexity of the segmentation algorithm. For each slide in the database, cell segmentation is performed using the apply_segmentation_with_download function, and the path where the result is saved is stored in the database using the load_seg_slide function. The segmentation process is similar to that described in the slide_cell_segmentation_with_download function, with the added step of downloading the segmented image and placing it in the same directory as the original slide. The segmentation process is defined in 6 steps:

LOAD SLIDE ... (slide_id)

APPLY SEGMENTATION ... (slide_id)

BUILD SEGMENTED SLIDE ... (slide_id)

BUILD GRAPH ... (slide_id)

BUILD & SAVE ADJACENCY MATRIX ... (slide_id)

SAVE SEGMENTED SLIDE ... (slide_id)

J-SPACE features ... (slide_id)

The adjacency matrix is saved in text format in the same directory as both the original and segmented images. Finally, a confirmation message is printed for each segmented slide.

# Examples with valid input
julia> JHistint.slide_cell_segmentation_with_download("acc")
julia> JHistint.slide_cell_segmentation_with_download("bLca")

# Examples with invalid input
julia> JHistint.slide_cell_segmentation_with_download("ac")
julia> JHistint.slide_cell_segmentation_with_download("")

Support Functions for Cell Segmentation (segmentationManager.jl)

JHistint.apply_segmentation_without_download — Method

apply_segmentation_without_download(slide_info::Tuple{String, Vector{UInt8}, String})

The function performs the segmentation of a histological image, generates its corresponding graph, and translates it into a symmetric adjacency matrix with only 0s and 1s.

Arguments

slide_info::Tuple{String, Vector{UInt8}, String}: A tuple containing the slide ID, the image obtained from the DB, and the path of the original image file.

Return values

filepath_matrix: The path where the adjacency matrix is stored in .txt format.

matrix: The adjacency matrix constructed from the segmentation.

Notes

The function uses the watershed segmentation algorithm to segment the image into different groups of pixels. Segmentation is performed using a feature transformation of the image (feature_transform) and labeling of connected components. The distance between the different regions is then calculated, and an adjacency graph of the regions is constructed using the region_adjacency_graph function. The resulting graph is then transformed into an adjacency matrix using the weighted_graph_to_adjacency_matrix function and saved to the path of the original image.

JHistint.apply_segmentation_with_download — Method

apply_segmentation_with_download(slide_info::Tuple{String, Vector{UInt8}, String})

The function performs segmentation of a histological image, saves the segmented image in .tif format, generates the corresponding graph, and translates it into an adjacency matrix.

Arguments

slide_info::Tuple{String, Vector{UInt8}, String}: Tuple containing the slide ID, the image itself obtained from the DB, and the path of the original image file.

Return values

filepath_seg: The path where the segmented image is stored in .tif format.

filepath_matrix: The path where the graph is stored in .txt format.

matrix: The adjacency matrix constructed from the segmentation.

Notes

The function uses the watershed segmentation algorithm to segment the image into different groups of pixels. Segmentation is performed using an image feature transformation (feature_transform) and connected component labeling. The distance between different regions is then calculated, and an adjacency graph of the regions is constructed using the region_adjacency_graph function. The obtained graph is transformed into an adjacency matrix using the weighted_graph_to_adjacency_matrix function, which is saved in the path of the original image. Finally, a segmented .tif image is saved, and the path of the segmented slide file is returned.

JHistint.save_adjacency_matrix — Method

save_adjacency_matrix(matrix::Matrix{Int64}, filepath_matrix::AbstractString)

Function to save an adjacency matrix represented as an integer matrix to a text file.

Arguments:

matrix::Matrix{Int64}: The integer matrix representing the adjacency matrix.

filepath_matrix::AbstractString: The file path represented as a string indicating where to save the matrix.

Notes

The function opens the file specified by the filepath_matrix path in write mode and writes the matrix in adjacency matrix format, where each row represents the adjacent nodes of a node. The numbers in the matrix are separated by spaces.

JHistint.weighted_graph_to_adjacency_matrix — Method

weighted_graph_to_adjacency_matrix(G::SimpleWeightedGraph{Int64, Float64}, n::Int64)

Converts a weighted graph represented as a SimpleWeightedGraph into an unweighted adjacency matrix.

Arguments:

G::SimpleWeightedGraph{Int64, Float64}: Weighted graph represented as a SimpleWeightedGraph with integer vertex labels and floating-point edge weights.

n::Int64: Number of nodes in the adjacency matrix.

Return value:

adjacency_matrix: Matrix{Int64} adjacency matrix.

Notes:

The function returns an n x n adjacency matrix representing the unweighted graph. If nodes i and j are adjacent, the adjacency matrix will contain a value of 1 at position (i,j) and (j,i). Otherwise, the adjacency matrix will contain a value of 0.

JHistint.get_random_color — Method

get_random_color(seed)

Function to return a random 8-bit RGB format color, using a specified seed.

Arguments

seed: An integer used to initialize the random number generator. If two calls to the function use the same seed, the same color will be generated.

Return value

The function returns a random 8-bit RGB format color.

DB Support Functions (dbManager.jl)

JHistint.insert_record_DB — Method

insert_record_DB(col_name::AbstractString,
                      cas_name::AbstractString,
                      tcga_case_id::AbstractString,
                      sin_cas_name::AbstractString,
                      tcga_slide_id::AbstractString,
                      link_slide::AbstractString,
                      filepath_zip::AbstractString,
                      filepath_svs::AbstractString)

Function for storing in the JHistint_DB database the information associated with each slide downloaded from the Cancer Digital Slide Archive (CDSA).

Argomenti

col_name::AbstractString = Collection name.

cas_name::AbstractString = Case name, which corresponds to the CASE-NAME displayed by the package.

tcga_case_id::AbstractString = ID used by TCGA to identify the case.

sin_cas_name::AbstractString = Name of the individual slide, which corresponds to the SLIDE-ID displayed by the package.

tcga_slide_id::AbstractString = ID used by TCGA to identify the slide.

link_slide::AbstractString = Link to the APIs for downloading the slide.

filepath_zip::AbstractString = Path where the .zip file is stored.

filepath_svs::AbstractString = Path where the .tif file is stored.

Notes

The JHistint_DB database is used for storing the information associated with each slide downloaded from the CDSA. The function takes a dictionary containing the information associated with the slide and stores it in the database. Data available in the JHistint_DB database for each slide:

collection_name TEXT = Name of the collection.

case_name TEXT = Name of the case.

TCGA_caseID TEXT = ID used by TCGA to identify the case.

slide_ID TEXT = Name of the individual slide case.

TCGA_slideID TEXT UNIQUE = ID used by TCGA to identify the slide, UNIQUE prevents duplicates from being generated.

slide_path_folder_zip TEXT = Path where the .zip file is stored.

slide_path_folder_svs TEXT = Path where the .tif file is stored.

slide_path_api TEXT = Link to the API for downloading the slide.

slide_path_folder_seg TEXT = Path where the segmented .tif file is stored.

slide_svs BLOB = Histopathological slide (image).

slide_info_TSS TEXT = Slide information - Tissue Source Site.

slide_info_participant_code TEXT = Slide information - Participant Code, alphanumeric string.

slide_info_sample_type TEXT = Slide information - Sample Type. The values associated with tumor samples are in the range 01-09. 10-19 indicates the range for non-diseased normal samples. 20-29 indicates samples currently under control.

slide_info_vial TEXT = Slide information - Vial. Related to the ordering of the sample in the sequence of samples. The values range from A-Z.

slide_info_portion TEXT = Slide information - Portion. Related to the ordering of the analyzed portions associated with a sample. Takes values in the range 01-99.

slide_info_type TEXT = Slide information - Image Type. The possible values are TS (Top Slide), BS (Bottom Slide), and MS (Middle Slide). The alphanumeric value indicates the ordering of the slide.

slide_path_folder_matrix TEXT = Path where the adjacency matrix .txt file is stored.

matrix_data BLOB = Adjacency matrix.

JHistint.query_extract_slide_svs — Method

query_extract_slide_svs(collection_name::AbstractString)

The function queries the JHistint_DB and extracts the list of slides associated with the collection name provided as an argument.

Arguments

collection_name::AbstractString: Name of the slide collection to search for in the JHistint_DB.

Return value

slide_list: List of tuples, each of which contains the ID of the slide, the .svs file of the slide, and the path of the folder containing the .svs file.

JHistint.load_seg_slide — Method

load_seg_slide(filepath_seg::AbstractString, filepath_matrix::AbstractString, matrix::Matrix{Int64}, slide_id::AbstractString)

The function updates the JHistint_DB with the path of the segmented image file, the path of the adjacency matrix file in text format, and the matrix itself.

Arguments

filepath_seg::AbstractString: Path of the segmented image file to add to the DB.

filepath_matrix::AbstractString: Path of the adjacency matrix file.

matrix::Matrix{Int64}: Adjacency matrix.

slide_id::AbstractString: ID of the slide to update with the segmented image information.

API Support Functions (apiManager.jl)

JHistint.download_collection_values — Method

download_collection_values(filepath::AbstractString)

Function for downloading data from collections available in TCGA.

Arguments

filepath::AbstractString = Path where to save the obtained .json file from the API available in CDSA.

Notes

The API requires the definition of parentType and parentId. parentId specifies the identifier of the collection. The collection of images associated with TCGA is identified by the code: 5b9ef8e3e62914002e454c39. The use of limit=0 sets the absence of limits in the queried file, ensuring the complete download of the file. The API belongs to the category for managing the folders stored in the repository. The downloaded file is .json.

JHistint.extract_collection_values — Method

extract_collection_values(filepath::AbstractString)

Function to extract the values of data collections from the .json file downloaded by the download_collection_values function.

Arguments

filepath::AbstractString = Path where the collectionlist.json file is stored.

Return value

collection_values::Array{String} = List of data collections available in TCGA.

JHistint.download_project_infos — Method

download_project_infos(filepath::AbstractString, collection_name::AbstractString)

Function to download metadata associated with the selected collection at startup.

Arguments

filepath::AbstractString = Path to save the .json file associated with the collection. The file is indicated

with the wording collection_name.json.

collection_name::AbstractString = Name of the collection from which to download the slides.

Notes

The API requires the definition of parentType, parentId, and name. The name attribute identifies the name of the collection from which you want to retrieve data (e.g., chol, esca, etc.). The API belongs to the category for managing the folders stored in the repository. The downloaded file is .json.

JHistint.extract_project_id — Method

extract_project_id(filepath::AbstractString)

Function to extract the id value from the metadata of the collection selected at startup.

Arguments

filepath::AbstractString = Path where the collection_name.json file is stored.

Return value

project_id = id of the collection.

JHistint.getCasesForProject — Method

getCasesForProject(filepath_case::AbstractString, project_id::AbstractString)

Function to download metadata associated with the cases of the selected collection at startup.

Arguments

filepath::AbstractString = Path where to save the .json file associated with the cases of the collection. The file is indicated with the term collection_name.json.

project_id::AbstractString = id of the collection.

Return values

casesID_values = List of id of all the cases in the collection.

casesNAME_values = List of name of all the cases in the collection.

Notes

The API requires the definition of parentType and parentId. The parentType attribute is set to folder given the structure of the repository. The parentId is set by defining the identifier of the chosen collection. The downloaded file is .json.

JHistint.download_zip — Method

download_zip(link::AbstractString, filepath::AbstractString)

Function for downloading histological slides in .zip format associated with the cases of the selected collection at startup.

Arguments

link::AbstractString = URL to access the API for slide download.

filepath::AbstractString = Path to save the .zip file.

ZIP Support Functions (zipManager.jl)

JHistint.extract_slide — Method

extract_slide(filepath_zip::AbstractString)

Function to extract the contents of .zip files downloaded from CDSA.

Arguments

filepath_zip::AbstractString = Path where the .zip file for the individual case is saved.